fastfetch: convert config to nix
The config was previously in `jsonc` format. This isn't
necessarily a problem, but `jsonc` is very obviously
a worse format than the `nix` DSL! (maybe)

At the least, this enabled changing a very, very long line
easily to an embedded shell script with embedded syntax
highlighting on at least `neovim` with `tree-sitter`.

This change also makes the embedded shell script an actual
real executable shell script that goes into the nix store.
It doesn't serve much value, but why not have a reusable
script when it's free as a result of other improvements?